[Fixes #26] Use BigDecimal for EMI interest rate to support rates like 12.5%, 13.5% etc.

+++ android-sdk/src/main/java/com/instamojo/android/fragments/EMIBankOptionsView.java
@@ -29,9 +29,9 @@ public class EMIBankOptionsView extends BaseFragment {
         return optionsView;
     }
 
-    private static double getEmiAmount(String totalAmount, int rate, int tenure) {
+    private static double getEmiAmount(String totalAmount, BigDecimal rate, int tenure) {
         double parsedAmount = Double.parseDouble(totalAmount);
-        double perRate = ((double) rate) / 1200;
+        double perRate = rate.doubleValue() / 1200;
         double emiAmount = parsedAmount * perRate / (1 - Math.pow((1 / (1 + perRate)), tenure));
         return getRoundedValue(emiAmount, 2);
     }
@@ -76,7 +76,7 @@ public class EMIBankOptionsView extends BaseFragment {
     private void loadOptions() {
         optionsContainer.removeAllViews();
         String orderAmount = parentActivity.getOrder().getAmount();
-        for (final Map.Entry<Integer, Integer> option : selectedBank.getRates().entrySet()) {
+        for (final Map.Entry<Integer, BigDecimal> option : selectedBank.getRates().entrySet()) {
             View optionView = LayoutInflater.from(getContext()).inflate(R.layout.emi_option_view,
                     optionsContainer, false);
             double emiAmount = getEmiAmount(orderAmount, option.getValue(), option.getKey());
